+++ MordoEngine/src/Water/WaterTileData.h
@@ -6,4 +6,6 @@ struct WaterTileData
 	glm::vec3 position;
 	glm::vec3 scale;
 	float yPos;
+	float waveStrength = 0.02f;
+	float moveFactor = 0.0f;
 };